requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
transformers==4.33.2
scikit-learn==1.3.0
python-magic==0.4.27
moviepy==1.0.3
librosa==0.10.1
soundfile==0.12.1
Werkzeug==2.3.7
gunicorn==21.2.0